Move copy-feature builders to helpers/copyFeatures.ts
Relocate buildRotatedCopies, buildMirroredCopies, buildCopiedFeatures,
buildCopiedClamps, and buildCopiedTabs to a dedicated home module.
pendingCompletionSlice now imports these directly.
