+++ src/store/projectStore.ts
@@ -39,7 +39,6 @@ import type {
   FeatureOperation,
   FeatureTreeEntry,
   Operation,
-  Point,
   Project,
   SketchProfile,
   SketchFeature,
@@ -69,21 +68,13 @@ import {
 } from './helpers/derivedFeatures'
 import { idNumericSuffix, nextUniqueGeneratedId, syncIdCounter } from './helpers/ids'
 import {
-  inferProfileOrientationAngle,
-  normalizeAngleDegrees,
   normalizeFeatureZRange,
   normalizeTool,
 } from './helpers/normalize'
 import {
-  rotatePointAround,
   transformProfile,
-  transformStlFeatureData,
-  translatePoint,
-  translateProfile,
 } from './helpers/transform'
-import { mirrorFeatureFromReference } from './helpers/referenceTransforms'
 import { isImportedModelFeature, normalizeImportedModelStorage, pruneUnusedModelAssets } from './helpers/modelAssets'
-import { duplicateClampName, duplicateFeatureName, duplicateTabName } from './helpers/naming'
 import { fallbackOperationTarget, defaultOperationForTarget, isOperationTargetValid } from './helpers/operationDefaults'
 import { createPendingAddSlice } from './slices/pendingAddSlice'
 import { createPendingActionsSlice } from './slices/pendingActionsSlice'
@@ -274,185 +265,6 @@ function clearStaleConstraints(features: SketchFeature[], movedIds: Set<string>)
     return feature
   })
   return anyChanged ? next : features
-}
-
-function buildRotatedCopies(
-  sourceFeatures: SketchFeature[],
-  existingFeatures: SketchFeature[],
-  pivot: Point,
-  angle: number,
-  count: number,
-): SketchFeature[] {
-  const created: SketchFeature[] = []
-  const projectLike: Project = { ...newProject(), features: existingFeatures, tools: [], operations: [] }
-
-  for (let step = 1; step <= count; step += 1) {
-    const stepAngle = angle * step
-    const rotatePoint = (point: Point) => rotatePointAround(point, pivot, stepAngle)
-    for (const sourceFeature of sourceFeatures) {
-      const nextId = nextUniqueGeneratedId(
-        { ...projectLike, features: [...existingFeatures, ...created] },
-        'f',
-      )
-      const profile = transformProfile(sourceFeature.sketch.profile, rotatePoint)
-      created.push({
-        ...sourceFeature,
-        id: nextId,
-        name: duplicateFeatureName(sourceFeature.name, [...existingFeatures, ...created], count, step),
-        folderId: sourceFeature.folderId,
-        stl: transformStlFeatureData(sourceFeature.stl, rotatePoint),
-        sketch: {
-          ...sourceFeature.sketch,
-          origin: rotatePoint(sourceFeature.sketch.origin),
-          orientationAngle: normalizeAngleDegrees(
-            (sourceFeature.sketch.orientationAngle ?? inferProfileOrientationAngle(sourceFeature.sketch.profile)) + stepAngle * (180 / Math.PI),
-          ),
-          profile,
-        },
-        locked: false,
-      })
-    }
-  }
-
-  return created
-}
-
-function buildMirroredCopies(
-  sourceFeatures: SketchFeature[],
-  existingFeatures: SketchFeature[],
-  lineStart: Point,
-  lineEnd: Point,
-): SketchFeature[] {
-  const created: SketchFeature[] = []
-  const projectLike: Project = { ...newProject(), features: existingFeatures, tools: [], operations: [] }
-
-  for (const sourceFeature of sourceFeatures) {
-    const nextId = nextUniqueGeneratedId(
-      { ...projectLike, features: [...existingFeatures, ...created] },
-      'f',
-    )
-    const mirrored = mirrorFeatureFromReference(sourceFeature, lineStart, lineEnd)
-    if (!mirrored) {
-      continue
-    }
-
-    created.push({
-      ...mirrored,
-      id: nextId,
-      name: duplicateFeatureName(sourceFeature.name, [...existingFeatures, ...created], 1, 1),
-      folderId: sourceFeature.folderId,
-      locked: false,
-    })
-  }
-
-  return created
-}
-
-function buildCopiedFeatures(
-  sourceFeatures: SketchFeature[],
-  existingFeatures: SketchFeature[],
-  dx: number,
-  dy: number,
-  count: number,
-): SketchFeature[] {
-  const created: SketchFeature[] = []
-  const projectLike: Project = {
-    ...newProject(),
-    features: existingFeatures,
-    tools: [],
-    operations: [],
-  }
-
-  for (let step = 1; step <= count; step += 1) {
-    for (const sourceFeature of sourceFeatures) {
-      const nextId = nextUniqueGeneratedId(
-        {
-          ...projectLike,
-          features: [...existingFeatures, ...created],
-        },
-        'f',
-      )
-      created.push({
-        ...sourceFeature,
-        id: nextId,
-        name: duplicateFeatureName(sourceFeature.name, [...existingFeatures, ...created], count, step),
-        folderId: sourceFeature.folderId,
-        stl: transformStlFeatureData(sourceFeature.stl, (point) => translatePoint(point, dx * step, dy * step)),
-        sketch: {
-          ...sourceFeature.sketch,
-          origin: ['text', 'stl'].includes(sourceFeature.kind)
-            ? { x: sourceFeature.sketch.origin.x + dx * step, y: sourceFeature.sketch.origin.y + dy * step }
-            : sourceFeature.sketch.origin,
-          profile: translateProfile(sourceFeature.sketch.profile, dx * step, dy * step),
-        },
-        locked: false,
-      })
-    }
-  }
-
-  return created
-}
-
-function buildCopiedClamps(
-  sourceClamps: Clamp[],
-  existingClamps: Clamp[],
-  project: Project,
-  dx: number,
-  dy: number,
-  count: number,
-): Clamp[] {
-  const created: Clamp[] = []
-
-  for (let step = 1; step <= count; step += 1) {
-    for (const sourceClamp of sourceClamps) {
-      created.push({
-        ...sourceClamp,
-        id: nextUniqueGeneratedId(
-          {
-            ...project,
-            clamps: [...existingClamps, ...created],
-          },
-          'cl',
-        ),
-        name: duplicateClampName(sourceClamp.name, [...existingClamps, ...created]),
-        x: sourceClamp.x + dx * step,
-        y: sourceClamp.y + dy * step,
-      })
-    }
-  }
-
-  return created
-}
-
-function buildCopiedTabs(
-  sourceTabs: Tab[],
-  existingTabs: Tab[],
-  project: Project,
-  dx: number,
-  dy: number,
-  count: number,
-): Tab[] {
-  const created: Tab[] = []
-
-  for (let step = 1; step <= count; step += 1) {
-    for (const sourceTab of sourceTabs) {
-      created.push({
-        ...sourceTab,
-        id: nextUniqueGeneratedId(
-          {
-            ...project,
-            tabs: [...existingTabs, ...created],
-          },
-          'tb',
-        ),
-        name: duplicateTabName(sourceTab.name, [...existingTabs, ...created]),
-        x: sourceTab.x + dx * step,
-        y: sourceTab.y + dy * step,
-      })
-    }
-  }
-
-  return created
 }
 
 export function syncFeatureTreeProject(project: Project): Project {
@@ -1081,11 +893,6 @@ export const useProjectStore = create<ProjectStore>((rawSet, get) => {
         return validateConstraintsOnFeature(f, byId)
       })
     },
-    buildCopiedFeatures,
-    buildCopiedClamps,
-    buildCopiedTabs,
-    buildRotatedCopies,
-    buildMirroredCopies,
     previewOffsetFeatures,
     syncFeatureTreeProject,
     createDerivedFeature,
